python/requirements.txt
python-dotenv
psycopg2
weaviate-client
flask
